+++ brainpy/connectivity/formater.py
