+++ brainpy/core_system/types.py
@@ -78,6 +78,9 @@ class ObjState(dict, TypeChecker):
     def __str__(self):
         return f'{self.__class__.__name__} ({str(self._keys)})'
 
+    def __repr__(self):
+        return self.__str__()
+
 
 class NeuState(ObjState):
     """Neuron State Management. """
