+++ brainpy/dynamics/analyst.py
