+++ brainpy/numpy/linalg.py
