+++ brainpy/numpy/random.py
