+++ brainpy/tools/ast2code.py
@@ -1084,3 +1084,5 @@ class Transformer(ast.NodeVisitor):
         self.write(node.name)
         if node.asname is not None:
             self.write(' as %s' % node.asname)
+
+
