+++ brainpy/tools/functions.py
@@ -5,9 +5,9 @@ import inspect
 import types
 
 from .codes import deindent
-from ..integration.integrator import Integrator
 from .. import numpy as np
 from .. import profile
+from ..integration.integrator import Integrator
 
 try:
     import numba as nb
@@ -28,8 +28,6 @@ __all__ = [
 def get_func_name(func, replace=False):
     func_name = func.__name__
     if replace:
-        func_name = func_name.replace('_npbrain_delay_push_', '')
-        func_name = func_name.replace('_npbrain_delay_pull_', '')
         func_name = func_name.replace('_npbrain_delayed_', '')
     return func_name
 
@@ -113,25 +111,65 @@ def numba_func(func, params={}):
         return jit(func)
 
 
+def _update_scope(k, v, scope):
+    if type(v).__name__ in ['module', 'function']:
+        return
+    if isinstance(v, Integrator):
+        return
+    if k in scope:
+        if v != scope[k]:
+            raise ValueError(f'Find scope variable {k} have different values: \n'
+                             f'{k} = {v} and {k} = {scope[k]}. \n'
+                             f'This maybe cause a grievous mistake in the future. Please change!')
+    scope[k] = v
+
+
 def get_func_scope(func, include_dispatcher=False):
+    """Get function scope variables.
+
+    Parameters
+    ----------
+    func : callable, Integrator
+    include_dispatcher
+
+    Returns
+    -------
+
+    """
+    # get function scope
     if isinstance(func, Integrator):
-        vars = inspect.getclosurevars(func.update_func)
+        func_name = func.py_func_name
+        variables = inspect.getclosurevars(func.update_func)
     elif type(func).__name__ == 'function':
-        vars = inspect.getclosurevars(func)
+        func_name = get_func_name(func, replace=True)
+        variables = inspect.getclosurevars(func)
     else:
         raise ValueError(f'Unknown type: {type(func)}')
-
-    scope = dict(vars.nonlocals)
-    scope.update(vars.globals)
+    scope = dict(variables.nonlocals)
+    scope.update(variables.globals)
 
     for k, v in list(scope.items()):
+        # get the scope of the function item
         if callable(v):
             if Dispatcher is not None and isinstance(v, Dispatcher):
                 if include_dispatcher:
-                    v_scope = get_func_scope(v.py_func)
-                    scope.update(v_scope)
+                    for k2, v2 in get_func_scope(v.py_func).items():
+                        try:
+                            _update_scope(k2, v2, scope)
+                        except ValueError:
+                            raise ValueError(f'Definition error in function "{func_name}".')
             else:
-                v_scope = get_func_scope(v)
-                scope.update(v_scope)
+                for k2, v2 in get_func_scope(v).items():
+                    try:
+                        _update_scope(k2, v2, scope)
+                    except ValueError:
+                        raise ValueError(f'Definition error in function "{func_name}".')
+
+    for k in list(scope.keys()):
+        v = scope[k]
+        if type(v).__name__ in ['module', 'function']:
+            scope.pop(k)
+        if isinstance(v, Integrator):
+            scope.pop(k)
 
     return scope
