+++ examples/synapses/AMPA_vector.py
@@ -2,8 +2,8 @@
 
 import matplotlib.pyplot as plt
 
-import npbrain as nb
-from npbrain import numpy as np
+import brainpy as nb
+from brainpy import numpy as np
 
 
 def AMPA1(g_max=0.10, E=0., tau_decay=2.0):
